Conflicts:
	Assets/Scripts/View/NoviceGuid/steps/NoviceGuideStepA.cs
	Assets/Scripts/View/UI/Loading/LoadingLogic.cs

+++ Assets/Scripts/View/NoviceGuid/steps/NoviceGuideStepA.cs
@@ -82,9 +82,8 @@ public class NoviceGuideStepA_StateOne:NoviceGuidState
 	{
 		LogHelper.Log("goto the selectrole");
 		//selectRole
-//		NoviceGuideStepEntityManager.FinishCurrentStep ();
-//		NoviceGuideStepEntityManager.CurrentNoviceGuideStage =  ;
-		NoviceGuideStepEntityManager.CurrentNoviceGuideStage = (int)NoviceGuideStage.SELECT_ROLE;
+		//NoviceGuideStepEntityManager.FinishCurrentStep ();
+		NoviceGuideStepEntityManager.CurrentNoviceGuideStage = NoviceGuideStage.SELECT_ROLE;
 		UIManager.Instance.ChangeScene (SceneEnum.SelectRole);
 	}
 }
